updates for task 5

+++ Task5/src/app/pages/courses/course-item/course-item.component.ts
@@ -19,7 +19,9 @@ export class CourseItemComponent implements OnChanges {
   }
 
   deleteCourse() {
-    this.deleteItem.emit(this.item);
+    if (confirm(`Do you really want to delete the course '${this.item.title}'?`)) {
+      this.deleteItem.emit(this.item);
+    }
   }
 
   editCourse() {
